+++ src/main/java/com/raoulvdberge/refinedstorage/container/ContainerBase.java
@@ -4,6 +4,7 @@ import com.raoulvdberge.refinedstorage.RSItems;
 import com.raoulvdberge.refinedstorage.apiimpl.API;
 import com.raoulvdberge.refinedstorage.container.slot.*;
 import com.raoulvdberge.refinedstorage.tile.TileBase;
+import invtweaks.api.container.InventoryContainer;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.ClickType;
 import net.minecraft.inventory.Container;
@@ -13,6 +14,7 @@ import net.minecraftforge.items.ItemHandlerHelper;
 
 import javax.annotation.Nonnull;
 
+@InventoryContainer(showOptions = false)
 public abstract class ContainerBase extends Container {
     private TileBase tile;
     private EntityPlayer player;