+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Refined Storage Changelog
 
+### 1.5.16
+- Fixed crash when placing a Controller (raoulvdberge)
+
 ### 1.5.15
 - Updated Forge to 2444 (MC 1.12.1) (raoulvdberge)
 - Added InventoryTweaks Grid sorting (cooliojazz)